requirements.txt
pyobjc-framework-EventKit
pyobjc-framework-Cocoa
PySide6
colorama
setproctitle
pytest
ruff
